feat: Link to author
Adds a new class, `BlogAuthor`. When creating a `BlogPost`, any class
that composes the `CreateBlogPostFromDataArray` trait will pull author
metadata from the appropriate YAML file and use it to create and compose
a `BlogAuthor` instance as a property of `BlogPost`. If no metadata is
found, it uses only the username, with the email address
no-reply@getlaminas.org.

Code for seeding the database and creating RSS feeds has been updated to
reflect the changes.